scMult0Int: 64bit

+++ kernel/combinatorics/stairc.h
@@ -25,7 +25,7 @@ int scDimInt(ideal  s,ideal Q=NULL); // dim (cf: field)
 /// scDimInt for ring-coefficients
 int scDimIntRing(ideal  s,ideal Q=NULL); // dim (general)
 int scMultInt(ideal  s,ideal Q=NULL); // mult
-int scMult0Int(ideal  s,ideal Q=NULL); // vdim
+long scMult0Int(ideal  s,ideal Q=NULL); // vdim
 void scPrintDegree(int co, int mu);
 void scDegree(ideal  s,intvec *modulweight,ideal Q=NULL);
